Introduce EmbeddingCenterContextBatchify API alike changed BPTT API
